- Create and Create list now returns the is of the created entities

+++ ums-api/src/main/java/org/ums/domain/model/immutable/SeatPlan.java
@@ -18,7 +18,7 @@ public interface SeatPlan extends Serializable, LastModifier, EditType<MutableSe
 
   Semester getSemester();
 
-  int getClassRoomId();
+  Long getClassRoomId();
 
   int getRowNo();
