+++ ums-api/src/main/java/org/ums/manager/UGTheoryMarksManager.java
@@ -4,5 +4,5 @@ import org.ums.domain.model.immutable.UGTheoryMarks;
 import org.ums.domain.model.mutable.MutableUGTheoryMarks;
 
 public interface UGTheoryMarksManager extends
-    ContentManager<UGTheoryMarks, MutableUGTheoryMarks, Integer> {
+    ContentManager<UGTheoryMarks, MutableUGTheoryMarks, Long> {
 }
