+++ ums-core/src/main/java/org/ums/decorator/ClassRoomDaoDecorator.java
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by Ifti on 13-Feb-16.
  */
 public class ClassRoomDaoDecorator extends
-    ContentDaoDecorator<ClassRoom, MutableClassRoom, Integer, ClassRoomManager> implements
+    ContentDaoDecorator<ClassRoom, MutableClassRoom, Long, ClassRoomManager> implements
     ClassRoomManager {
   @Override
   public List<ClassRoom> getRoomList() {
